+++ CSharpGL/CSharpGL.csproj
@@ -94,6 +94,7 @@
     <Compile Include="FontResource\FontResource.FromTTF.cs" />
     <Compile Include="NewFontResource\FontBitmap.cs" />
     <Compile Include="NewFontResource\FontBitmapHelper.cs" />
+    <Compile Include="NewFontResource\FontTexture.IDisposable.cs" />
     <Compile Include="NewFontResource\FontTexture.cs" />
     <Compile Include="NewFontResource\FontTextureHelper.cs" />
     <Compile Include="Renderers\IBufferables\BoundingBoxModel.cs" />
