+++ CSharpGL/NewFontResource/FontTexture.IDisposable.cs
@@ -1,0 +1,53 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+
+namespace CSharpGL.NewFontResource
+{
+    /// <summary>
+    /// font, texture and texture coordiante.
+    /// </summary>
+    public partial class FontTexture
+    {
+
+        /// <summary>
+        /// 
+        /// </summary>
+        public void Dispose()
+        {
+            this.Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        ~FontTexture()
+        {
+            this.Dispose(false);
+        }
+
+        private bool disposedValue = false;
+
+        private void Dispose(bool disposing)
+        {
+            if (this.disposedValue == false)
+            {
+                if (disposing)
+                {
+                    // Dispose managed resources.
+
+                }
+
+                // Dispose unmanaged resources.
+                var ids = new uint[] { FontTextureId, };
+                OpenGL.DeleteTextures(ids.Length, ids);
+            }
+
+            this.disposedValue = true;
+        }
+
+    }
+}
